+++ support/__init__.py
@@ -1,13 +1,13 @@
 '''
-File: __init__.py
-Project: edutours.com.ng
+File: user.py
+Project: token-credit-backend
 File Created: Sunday, 26th January 2020 12:34:04 am
-Author: Temitayo Bodunrin (temitayo@brandnaware.com)
+Author: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Last Modified: Wednesday, 26th February 2020 7:24:19 pm
-Modified By: Temitayo Bodunrin (temitayo@brandnaware.com)
+Last Modified: Wednesday, 28th October 2020 1:01:01 pm
+Modified By: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Copyright 2020, Brandnaware Nigeria
+Copyright 2020, CamelCase Technologies Ltd
 '''
 # from . import http
 # from . import helper