+++ support/auth/backends.py
@@ -1,13 +1,13 @@
 '''
-File: backends.py
-Project: edutours.com.ng
+File: user.py
+Project: token-credit-backend
 File Created: Wednesday, 19th February 2020 5:09:55 pm
-Author: Temitayo Bodunrin (temitayo@brandnaware.com)
+Author: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Last Modified: Wednesday, 19th February 2020 5:59:47 pm
-Modified By: Temitayo Bodunrin (temitayo@brandnaware.com)
+Last Modified: Wednesday, 28th October 2020 1:01:01 pm
+Modified By: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Copyright 2020, Brandnaware Nigeria
+Copyright 2020, CamelCase Technologies Ltd
 '''
 from django.contrib.auth.backends import BaseBackend
 from django.apps import apps as django_apps