+++ support/storages/default_backend.py
@@ -1,13 +1,13 @@
 '''
-File: default_backend.py
-Project: edutours.com.ng
+File: user.py
+Project: token-credit-backend
 File Created: Thursday, 20th February 2020 3:09:51 pm
-Author: Temitayo Bodunrin (temitayo@brandnaware.com)
+Author: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Last Modified: Thursday, 20th February 2020 4:13:41 pm
-Modified By: Temitayo Bodunrin (temitayo@brandnaware.com)
+Last Modified: Wednesday, 28th October 2020 1:01:02 pm
+Modified By: Temitayo Bodunrin (temitayo@camelcase.co)
 -----
-Copyright 2020, Brandnaware Nigeria
+Copyright 2020, CamelCase Technologies Ltd
 '''
 
 from ..helper import config, getClassFromDotString